requirements.txt
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cmake==4.1.2
colorlog==6.10.1
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.2
cycler==0.12.1
dash==3.3.0
decorator==5.2.1
einops==0.7.0
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
Flask==3.1.2
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.2.0
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.0
ipython==8.37.0
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lightning-utilities==0.15.2
lit==18.1.8
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mpmath==1.3.0
multidict==6.7.0
narwhals==2.12.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.25.2
omegaconf==2.3.0
open3d==0.18.0
orjson==3.9.7
packaging==25.0
pandas==2.0.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
plotly==6.5.0
plyfile==1.1.3
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytorch-lightning==2.0.9
pytz==2025.2
PyYAML==6.0.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
retrying==1.4.2
rpds-py==0.29.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.11.1
sentry-sdk==2.45.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.0.1+cu118
torchmetrics==1.8.2
tqdm==4.65.0
traitlets==5.14.3
transformers==4.31.0
trimesh==4.4.1
triton==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.23.0
wcwidth==0.2.14
Werkzeug==3.1.3
widgetsnbextension==4.0.15
yarl==1.22.0
zipp==3.23.0
